Extract topic store subscription
